Rename some camel-case scripts in setup.py to a smqtk-... format
It turns out that pip install'ing and then uninstalling leaves around
console scripts that were spelled with a camel-case format. Not sure if
this is just my system or not, but it makes a little more sense to use a
"smqtk-..." prefix for the convenience of being able to use tab-complete
to see what CLI tools are available via smqtk.

Did not change all scripts since some are used in docker images which
would necessitate a larger change.