Major overhaul of tree construction
1. Originally, as in the paper, we were using sparse projection
matrices. Upon examination, this was unnecessary theoretically (from a
Johnson-Lindenstrauss standpoint) and practically (the projection matrix
is one of the smallest blocks of memory used by the algorithm).
Performing a sparse x dense multiplication where the sparse matrix was
small and the dense matrix was large took more time by a large factor.
This has been changed to use dense projection matrices and build time is
much faster.
2. As a result of (1) above, we no longer need to be careful about the
order in which we multiply matrices, so we adopt matrix dimensions that
are more intuitive. This also results in a data matrix that we don't
need to transpose, so this is an added speed advantage.
3. Projection matrices are filled in a block once, rather than
individually per-tree. The projections are also performed at once for
all trees.
4. The single-tree construction routine has been modified to take
advantage of the above modifications, as well as to use an
internally-scoped function to do the heavy lifting and recursion. This
keeps the variables that do not change in the outer scope, making
cleaner, more intuitive code in the recursive portion. (In fact, most of
the changes in _build_single_tree are whitespace changes or comment
changes.)