Revise error raised when no from_uri implementation
Raising a NotImplementedError causes the static analyzer to interpret
the container class as abstract (even without the use of the abc
module).  Changing the expected error to be thrown to a custom one
causes interface implementing classes that purposefully do not implement
from_uri from being labeled an abstract class and missing required
method implementations.