Implementing update and updating tests
This required some refactoring of the internals to make things flow a
little better. Since build and update share so much code, and little of
what's shared needs to be locked behind the mutex, move the common code
into its own function and then feed the results of that function to the
build or update logic.

Also, when building, build the new faiss index outside of the lock and
then add it to the object once it's built and we hold the lock, along
with clearing and filling the descriptor set and the uuid list.