Update documentation python deps, remove mocking from conf.py
The mocking that occurred in the ``docs/conf.py`` is brittle to changes
in the python code when new sub-modules are imported and only adds the
ability to build the documentation without an install of the SMQTK
python package and its code dependencies.  This mocking was previously
required because ReadTheDocs had previously only installed the
requirements from the specificly configured requirements file.  Now,
however, the whole SMQTK package is being installed via a ``python
setup.py install --force`` command, thus installing all of SMQTK's
core dependencies in the process of building the docs.